manager: Better integration of KPM in KernelSU manager
Much of the code was rewritten for better integration with KernelSU, including functions, declarations, etc.
The spacing and alignment issues that caused the entire KPM interface to be misaligned and even off-screen were also resolved.
MaterialDesign 3 Expressive was integrated into the KPM section to ensure visual consistency with the rest of the manager.
The KPM section was relocated and given a more appropriate icon to avoid confusion.

Signed-off-by: maxsteeel <quinteroalejandro2009@gmail.com>